[front] refactor(mcp): migrate run_dust_app server to new pattern
